Finish moving api.js to background code
